ECS integration (#190)
* Moved ECS work to plugins, and hooked it up so that it can be reported on chalk, extraction, etc.  Can still call readECSMetadata() directly to get a JSON object, and will only ever hit the IP once.

* Non-ECS related; reconfiging to make the logging more silent for docker, which included moving when some warnings get output to after the config files load.

+++ src/plugins/codecDocker.nim
@@ -590,6 +590,7 @@ proc buildContainer*(chalk:  ChalkObj,
 proc cleanupTmpFiles*(chalk: ChalkObj) =
   let cache = DockerInfoCache(chalk.cache)
 
+  if cache == nil:              return
   if cache.tmpDockerFile != "": removeFile(cache.tmpDockerFile)
   if cache.tmpChalkMark  != "": removeFile(cache.tmpChalkMark)
   if cache.tmpEntryPoint != "": removeFile(cache.tmpEntryPoint)
